some updates
- components separated
- Gulp sass compiler task added
