+++ README.md
@@ -1,2 +1,22 @@
 # react-hangman
-Hangman game with react 
+Hangman game with react
+
+#### CodePen
+You can also find it <a href="http://codepen.io/ozluy/pen/QEbgzx" >here </a> but there are bugs with Codepen according to **hotkey**
+
+#### Technologies
+- React
+- SASS(on CodePen page) / SCSS (local)
+- webpack
+- gulp
+
+# guide
+
+#### download node_modules
+`npm install`
+
+#### sass compile
+`gulp`
+
+#### run project
+`npm start`
