+++ package.json
@@ -10,6 +10,8 @@
   "license": "ISC",
   "dependencies": {
     "axios": "^0.12.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-sass": "^2.3.1",
     "react": "^15.1.0",
     "react-dom": "^15.1.0",
     "react-hotkey": "^0.6.0"
@@ -19,6 +21,8 @@
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.5.0",
+    "gulp": "^3.9.1",
+    "gulp-cssmin": "^0.1.7",
     "webpack": "^1.13.1",
     "webpack-dev-server": "^1.14.1"
   }
