--- src/bundle.css
@@ -1,294 +1,0 @@
-body {
-  font-weight: bold;
-  font-size: 14px;
-  background-color: #3C3F64;
-  padding: 50px;
-  font-family: sans-serif;
-}
-
-#app {
-  position: relative;
-  overflow: hidden;
-  height: 500px;
-  width: 900px;
-  background-color: #F5F5F5;
-  margin: 0 auto;
-  border-radius: 5px;
-}
-
-.gallow {
-  position: relative;
-  width: 150px;
-  height: 15px;
-  background-color: #53555D;
-  top: 50px;
-  border-radius: 0 5px 5px 0;
-}
-.gallow .down-pipe {
-  position: absolute;
-  left: 120px;
-  width: 15px;
-  height: 40px;
-  background-color: #53555D;
-  border-radius: 5px;
-}
-
-.man {
-  position: relative;
-  margin: 70px 0 0 77.5px;
-  width: 100px;
-}
-.man .head {
-  position: absolute;
-  z-index: 2;
-  margin: 0 auto;
-  top: 0;
-  left: 0;
-  right: 0;
-  width: 96px;
-  height: 106px;
-  background-image: url("https://raw.githubusercontent.com/ozluy/ozluy.github.io/master/source/head.png");
-  background-size: 96px 106px;
-  background-repeat: no-repeat;
-  background-position: center center;
-}
-.man .neck {
-  position: absolute;
-  margin: 0 auto;
-  top: 100px;
-  left: 0;
-  right: 0;
-  width: 20px;
-  height: 40px;
-  background-color: #FFCD82;
-}
-.man .corpus {
-  position: absolute;
-  margin: 0 auto;
-  top: 112.5px;
-  left: 0;
-  right: 0;
-  width: 60px;
-  height: 95px;
-  border-radius: 5px;
-  background-color: #3C3F64;
-}
-.man .corpus .corpus-chest-left {
-  display: inline-block;
-  width: 30px;
-  height: 70px;
-  background-color: #71CC54;
-  border-radius: 5px 0 0 0;
-}
-.man .corpus .corpus-chest-left .left-arm {
-  position: relative;
-  -webkit-transform: rotate(45deg);
-  transform: rotate(45deg);
-  margin-left: -27px;
-  margin-top: -5px;
-  width: 22px;
-  height: 80px;
-  background-color: #71CC54;
-  border-radius: 5px;
-}
-.man .corpus .corpus-chest-left .left-arm .left-hand {
-  position: absolute;
-  bottom: -14px;
-  border-radius: 0 0 50% 50%;
-  left: 0;
-  right: 0;
-  margin: 0 auto;
-  width: 15px;
-  height: 15px;
-  background-color: #FFCD82;
-}
-.man .corpus .corpus-chest-left .left-leg {
-  background-color: #3C3F64;
-  -webkit-transform: rotate(15deg);
-  transform: rotate(15deg);
-  height: 70px;
-  width: 25px;
-  margin-top: 10px;
-  margin-left: -10px;
-}
-.man .corpus .corpus-chest-left .left-leg .left-foot {
-  position: absolute;
-  bottom: -14px;
-  border-radius: 0 0 50% 50%;
-  left: -15px;
-  right: 0;
-  margin: 0 auto;
-  width: 40px;
-  height: 15px;
-  background-color: #FFCD82;
-}
-.man .corpus .corpus-chest-right {
-  display: inline-block;
-  width: 30px;
-  height: 70px;
-  background-color: #12A56B;
-  border-radius: 0 5px 0 0;
-}
-.man .corpus .corpus-chest-right .right-arm {
-  position: relative;
-  -webkit-transform: rotate(-45deg);
-  transform: rotate(-45deg);
-  margin-left: 35px;
-  margin-top: -5px;
-  width: 22px;
-  height: 80px;
-  background-color: #12A56B;
-  border-radius: 5px;
-}
-.man .corpus .corpus-chest-right .right-arm .right-hand {
-  position: absolute;
-  bottom: -14px;
-  border-radius: 0 0 50% 50%;
-  left: 0;
-  right: 0;
-  margin: 0 auto;
-  width: 15px;
-  height: 15px;
-  background-color: #FFCD82;
-}
-.man .corpus .corpus-chest-right .right-leg {
-  background-color: #3C3F64;
-  -webkit-transform: rotate(-15deg);
-  transform: rotate(-15deg);
-  height: 70px;
-  width: 25px;
-  margin-top: 10px;
-  margin-left: 15px;
-}
-.man .corpus .corpus-chest-right .right-leg .right-foot {
-  position: absolute;
-  bottom: -14px;
-  border-radius: 0 0 50% 50%;
-  left: 0;
-  right: 0;
-  margin: 0 auto;
-  width: 40px;
-  height: 15px;
-  background-color: #FFCD82;
-}
-
-.right-blue-triangle {
-  position: absolute;
-  width: 0;
-  height: 0;
-  right: 0;
-  bottom: 0;
-  border-style: solid;
-  border-width: 0 0 250px 250px;
-  border-color: transparent transparent #585DFE transparent;
-}
-
-.answer-box-menu {
-  z-index: 3;
-  position: absolute;
-  bottom: 40px;
-  width: calc(100% - 50px);
-  height: 70px;
-  margin: 0 auto;
-  text-align: center;
-  left: 0;
-  right: 0;
-}
-.answer-box-menu .answer-box-menu-item {
-  float: left;
-  background-color: #53555D;
-  width: 65px;
-  height: 70px;
-  margin: 0 3px;
-  border-radius: 5px;
-  text-align: center;
-  line-height: 70px;
-  text-transform: uppercase;
-  font-size: 36px;
-  color: #FFFFFF;
-}
-.answer-box-menu .answer-box-menu-item.disabled {
-  background-color: #E6E6E7;
-}
-.answer-box-menu .answer-box-menu-item:first-child {
-  margin-left: 0;
-}
-.answer-box-menu .answer-box-menu-item:last-child {
-  margin-right: 0;
-}
-
-.fail-box {
-  position: absolute;
-  top: 40px;
-  right: 0;
-  width: 500px;
-}
-.fail-box .fail-box-title {
-  text-transform: uppercase;
-  color: #53555D;
-}
-.fail-box .fail-box-menu {
-  width: 100%;
-  height: 70px;
-}
-.fail-box .fail-box-menu .fail-box-menu-item {
-  text-align: center;
-  color: #585DFE;
-  float: left;
-  font-size: 36px;
-  height: 70px;
-  margin: 10px;
-  text-transform: uppercase;
-}
-.fail-box .fail-box-menu .fail-box-menu-item:first-child {
-  margin-left: 0;
-}
-.fail-box .fail-box-menu .fail-box-menu-item:last-child {
-  margin-right: 0;
-}
-
-#game-result {
-  z-index: 10;
-  position: absolute;
-  background-color: rgba(0, 0, 0, 0.7);
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-}
-#game-result.disabled {
-  visiblity: 0;
-  opacity: 0;
-  z-index: -1;
-}
-#game-result .content {
-  top: calc(50% - 75px);
-  left: 0;
-  right: 0;
-  bottom: 0;
-  position: absolute;
-  margin: 0 auto;
-  width: 300px;
-  height: 150px;
-  text-align: center;
-}
-#game-result .content .result-title {
-  text-transform: uppercase;
-  color: #FFFFFF;
-  font-size: 36px;
-}
-#game-result .content .restart {
-  text-transform: uppercase;
-  border: dashed 2px #FFB800;
-  border-radius: 25px;
-  background-color: transparent;
-  color: #FFB800;
-  padding: 15px 40px;
-  font-size: 14px;
-  font-weight: bold;
-  outline: none;
-  cursor: pointer;
-}
-#game-result .content .restart:hover {
-  opacity: 0.6;
-}
