+++ src/components/FailBoxMenu.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+export default class FailBoxMenu  extends React.Component {
+  render(){
+    return (
+      <div className="fail-box">
+        <div className="fail-box-title">You missed: </div>
+        <div className="fail-box-menu">
+          {
+            this.props.children
+          }
+        </div>
+      </div>
+    );
+  }
+}
+FailBoxMenu.propTypes = {
+  children: React.PropTypes.node
+}
