+++ src/components/Gallow.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+export default class Gallow extends React.Component {
+  render(){
+    return (
+      <div className="gallow">
+        <div className="down-pipe"> </div>
+      </div>
+    );
+  }
+}
