+++ src/components/RightBlueTriangle.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+export default class RightBlueTriangle extends React.Component {
+  render(){
+    return (
+      <div className="right-blue-triangle">
+      </div>
+    );
+  }
+}
