+++ src/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="utf-8" />
     <title>Hangman || Adam Asmaca</title>
-    <link href="bundle.css" rel="stylesheet" type="text/css">
+    <link href="bundle.min.css" rel="stylesheet" type="text/css">
 </head>
 <body>
 <div id="app"></div>
