+++ src/styles.scss
@@ -1,0 +1,376 @@
+// Resources
+$head: "https://raw.githubusercontent.com/ozluy/ozluy.github.io/master/source/head.png";
+
+//Mixins
+@mixin rotate($val) {
+  -moz-transform: rotate($val);
+  -webkit-transform: rotate($val);
+  -o-transform: rotate($val);
+  -ms-transform: rotate($val);
+  transform: rotate($val);
+}
+
+@mixin calc($property, $val) {
+  #{$property}: -webkit-calc(#{$val});
+  #{$property}: -moz-calc(#{$val});
+  #{$property}: calc(#{$val});
+}
+
+@mixin transition($val) {
+  -webkit-transition: $val;
+  -moz-transition: $val;
+  -ms-transition: $val;
+  -o-transition: $val;
+  transition: $val;
+}
+
+// Colors *** Watch out! I am a color blind. ***
+$bodybg: #3C3F64;
+$blue: #585DFE;
+$darkgrey: #53555D;
+$lightgrey: #E6E6E7;
+$humanbody: #FFCD82;
+$shirtLeft: #71CC54;
+$shirtRight: #12A56B;
+$appbg: #F5F5F5;
+$white: #FFFFFF;
+$yellow: #FFB800;
+
+// Font Size
+$default: 14px;
+$large: 36px;
+
+// Font Weight
+$bold: bold;
+
+body {
+  font-weight: $bold;
+  font-size: $default;
+  background-color: $bodybg;
+  padding: 50px;
+  font-family: sans-serif;
+}
+
+#app {
+  position: relative;
+  overflow: hidden;
+  height: 500px;
+  width: 900px;
+  background-color: $appbg;
+  margin: 0 auto;
+  border-radius: 5px;
+}
+
+.gallow {
+  position: relative;
+  width: 150px;
+  height: 15px;
+  background-color: $darkgrey;
+  top: 50px;
+  border-radius: 0 5px 5px 0;
+
+  .down-pipe {
+    position: absolute;
+    left: 120px;
+    width: 15px;
+    height: 40px;
+    background-color: $darkgrey;
+    border-radius: 5px;
+  }
+}
+
+.man {
+  position: relative;
+  margin: 70px 0 0 77.5px;
+  width: 100px;
+
+  .head {
+    position: absolute;
+    z-index: 2;
+    margin: 0 auto;
+    top: 0;
+    left: 0;
+    right: 0;
+    width: 96px;
+    height: 106px;
+    background-image: url($head);
+    background-size: 96px 106px;
+    background-repeat: no-repeat;
+    background-position: center center;
+  }
+
+  .neck {
+    position: absolute;
+    margin: 0 auto;
+    top: 100px;
+    left: 0;
+    right: 0;
+    width: 20px;
+    height: 40px;
+    background-color: $humanbody;
+  }
+
+  .corpus {
+    position: absolute;
+    margin: 0 auto;
+    top: 112.5px;
+    left: 0;
+    right: 0;
+    width: 60px;
+    height: 95px;
+    border-radius: 5px;
+    background-color: $bodybg;
+
+    .corpus-chest-left {
+      display: inline-block;
+      width: 30px;
+      height: 70px;
+      background-color: $shirtLeft;
+      border-radius: 5px 0 0 0;
+
+      .left-arm {
+        position: relative;
+
+        @include rotate(45deg);
+
+        margin-left: -27px;
+        margin-top: -5px;
+        width: 22px;
+        height: 80px;
+        background-color: $shirtLeft;
+        border-radius: 5px;
+
+        .left-hand {
+          position: absolute;
+          bottom: -14px;
+          border-radius: 0 0 50% 50%;
+          left: 0;
+          right: 0;
+          margin: 0 auto;
+          width: 15px;
+          height: 15px;
+          background-color: $humanbody;
+        }
+      }
+
+      .left-leg {
+        background-color: $bodybg;
+
+        @include rotate(15deg);
+
+        height: 70px;
+        width: 25px;
+        margin-top: 10px;
+        margin-left: -10px;
+
+        .left-foot {
+          position: absolute;
+          bottom: -14px;
+          border-radius: 0 0 50% 50%;
+          left: -15px;
+          right: 0;
+          margin: 0 auto;
+          width: 40px;
+          height: 15px;
+          background-color: $humanbody;
+        }
+      }
+    }
+
+    .corpus-chest-right {
+      display: inline-block;
+      width: 30px;
+      height: 70px;
+      background-color: $shirtRight;
+      border-radius: 0 5px 0 0;
+
+      .right-arm {
+        position: relative;
+
+        @include rotate(-45deg);
+
+        margin-left: 35px;
+        margin-top: -5px;
+        width: 22px;
+        height: 80px;
+        background-color: $shirtRight;
+        border-radius: 5px;
+
+        .right-hand {
+          position: absolute;
+          bottom: -14px;
+          border-radius: 0 0 50% 50%;
+          left: 0;
+          right: 0;
+          margin: 0 auto;
+          width: 15px;
+          height: 15px;
+          background-color: $humanbody;
+        }
+      }
+
+      .right-leg {
+        background-color: $bodybg;
+
+        @include rotate(-15deg);
+
+        height: 70px;
+        width: 25px;
+        margin-top: 10px;
+        margin-left: 15px;
+
+        .right-foot {
+          position: absolute;
+          bottom: -14px;
+          border-radius: 0 0 50% 50%;
+          left: 0;
+          right: 0;
+          margin: 0 auto;
+          width: 40px;
+          height: 15px;
+          background-color: $humanbody;
+        }
+      }
+    }
+  }
+}
+
+.right-blue-triangle {
+  position: absolute;
+  width: 0;
+  height: 0;
+  right: 0;
+  bottom: 0;
+  border-style: solid;
+  border-width: 0 0 250px 250px;
+  border-color: transparent transparent $blue transparent;
+}
+
+.answer-box-menu {
+  z-index: 3;
+  position: absolute;
+  bottom: 40px;
+
+  @include calc(width, "100% - 50px");
+
+  height: 70px;
+  margin: 0 auto;
+  text-align: center;
+  left: 0;
+  right: 0;
+
+  .answer-box-menu-item {
+    float: left;
+    background-color: $darkgrey;
+    width: 65px;
+    height: 70px;
+    margin: 0 3px;
+    border-radius: 5px;
+    text-align: center;
+    line-height: 70px;
+    text-transform: uppercase;
+    font-size: $large;
+    color: $white;
+
+    &.disabled {
+      background-color: $lightgrey;
+    }
+
+    &:first-child {
+      margin-left: 0;
+    }
+
+    &:last-child {
+      margin-right: 0;
+    }
+  }
+}
+
+.fail-box {
+  position: absolute;
+  top: 40px;
+  right: 0;
+  width: 500px;
+
+  .fail-box-title {
+    text-transform: uppercase;
+    color: $darkgrey;
+  }
+
+  .fail-box-menu {
+    width: 100%;
+    height: 70px;
+
+    .fail-box-menu-item {
+      text-align: center;
+      color: $blue;
+      float: left;
+      font-size: $large;
+      height: 70px;
+      margin: 10px;
+      text-transform: uppercase;
+
+      &:first-child {
+        margin-left: 0;
+      }
+
+      &:last-child {
+        margin-right: 0;
+      }
+    }
+  }
+}
+
+#game-result {
+  z-index: 10;
+  position: absolute;
+  background-color: rgba(0, 0, 0, 0.7);
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+
+  @include transition(all 400ms ease-in-out);
+
+  &.disabled {
+    visiblity: 0;
+    opacity: 0;
+    z-index: -1;
+  }
+
+  .content {
+    @include calc(top, "50% - 75px");
+
+    left: 0;
+    right: 0;
+    bottom: 0;
+    position: absolute;
+    margin: 0 auto;
+    width: 300px;
+    height: 150px;
+    text-align: center;
+
+    .result-title {
+      text-transform: uppercase;
+      color: $white;
+      font-size: $large;
+    }
+
+    .restart {
+      text-transform: uppercase;
+      border: dashed 2px $yellow;
+      border-radius: 25px;
+      background-color: transparent;
+      color: $yellow;
+      padding: 15px 40px;
+      font-size: $default;
+      font-weight: $bold;
+      outline: none;
+      cursor: pointer;
+
+      &:hover {
+        opacity: 0.6;
+      }
+    }
+  }
+}
